+++ Packs/Base/Scripts/CommonServerPython/CommonServerPython_test.py
@@ -8,12 +8,13 @@ import sys
 import requests
 from pytest import raises, mark
 import pytest
+
 from CommonServerPython import xml2json, json2xml, entryTypes, formats, tableToMarkdown, underscoreToCamelCase, \
     flattenCell, date_to_timestamp, datetime, camelize, pascalToSpace, argToList, \
     remove_nulls_from_dictionary, is_error, get_error, hash_djb2, fileResult, is_ip_valid, get_demisto_version, \
     IntegrationLogger, parse_date_string, IS_PY3, DebugLogger, b64_encode, parse_date_range, return_outputs, \
     argToBoolean, ipv4Regex, ipv4cidrRegex, ipv6cidrRegex, ipv6Regex, batch, FeedIndicatorType, \
-    encode_string_results, safe_load_json, remove_empty_elements, aws_table_to_markdown
+    encode_string_results, safe_load_json, remove_empty_elements, aws_table_to_markdown, is_demisto_version_ge
 
 try:
     from StringIO import StringIO
@@ -34,6 +35,16 @@ INFO = {'b': 1,
             'c': {'d': 10},
         }
         }
+
+
+@pytest.fixture()
+def clear_version_cache():
+    """
+    Clear the version cache at end of the test (in case we mocked demisto.serverVersion)
+    """
+    yield
+    if hasattr(get_demisto_version, '_version'):
+        delattr(get_demisto_version, '_version')
 
 
 def test_xml():
@@ -735,7 +746,7 @@ def test_exception_in_return_error(mocker):
     assert IntegrationLogger.__call__.call_count == 2
 
 
-def test_get_demisto_version(mocker):
+def test_get_demisto_version(mocker, clear_version_cache):
     # verify expected server version and build returned in case Demisto class has attribute demistoVersion
     mocker.patch.object(
         demisto,
@@ -749,6 +760,26 @@ def test_get_demisto_version(mocker):
         'version': '5.0.0',
         'buildNumber': '50000'
     }
+    # call again to check cache
+    assert get_demisto_version() == {
+        'version': '5.0.0',
+        'buildNumber': '50000'
+    }
+    # call count should be 1 as we cached
+    assert demisto.demistoVersion.call_count == 1
+    # test is_demisto_version_ge
+    assert is_demisto_version_ge('5.0.0')
+    assert is_demisto_version_ge('4.5.0')
+    assert not is_demisto_version_ge('5.5.0')
+
+
+def test_is_demisto_version_ge_4_5(mocker):
+    get_version_patch = mocker.patch('CommonServerPython.get_demisto_version')
+    get_version_patch.side_effect = AttributeError('simulate missing demistoVersion')
+    assert not is_demisto_version_ge('5.0.0')
+    assert not is_demisto_version_ge('6.0.0')
+    with raises(AttributeError, match='simulate missing demistoVersion'):
+        is_demisto_version_ge('4.5.0')
 
 
 def test_assign_params():
@@ -836,6 +867,261 @@ class TestBuildDBotEntry(object):
         }
 
 
+class TestCommandResults:
+    def test_return_command_results(self):
+        from CommonServerPython import Common, CommandResults, EntryFormat, EntryType, DBotScoreType
+        ip = Common.IP(
+            ip='8.8.8.8',
+            asn='some asn',
+            hostname='test.com',
+            geo_country=None,
+            geo_description=None,
+            geo_latitude=None,
+            geo_longitude=None,
+            positive_engines=None,
+            detection_engines=None
+        )
+
+        dbot_score = Common.DBotScore(
+            indicator='8.8.8.8',
+            integration_name='Virus Total',
+            indicator_type=DBotScoreType.IP,
+            score=Common.DBotScore.GOOD
+        )
+        ip.set_dbot_score(dbot_score)
+
+        results = CommandResults(
+            outputs_key_field=None,
+            outputs_prefix=None,
+            outputs=None,
+            indicators=[ip]
+        )
+
+        assert results.to_context() == {
+            'Type': EntryType.NOTE,
+            'ContentsFormat': EntryFormat.JSON,
+            'Contents': None,
+            'HumanReadable': None,
+            'EntryContext': {
+                'IP(val.Address && val.Address == obj.Address)': {
+                    'Address': '8.8.8.8',
+                    'ASN': 'some asn',
+                    'Hostname': 'test.com'
+                },
+                'DBotScore(val.Indicator && val.Indicator == obj.Indicator && '
+                'val.Vendor == obj.Vendor && val.Type == obj.Type)': {
+                    'Indicator': '8.8.8.8',
+                    'Vendor': 'Virus Total',
+                    'Score': 1,
+                    'Type': 'ip'
+                }
+            }
+        }
+
+    def test_create_dbot_score(self):
+        from CommonServerPython import Common, CommandResults, EntryFormat, EntryType, DBotScoreType
+
+        dbot_score = Common.DBotScore(
+            indicator='8.8.8.8',
+            integration_name='Virus Total',
+            indicator_type=DBotScoreType.IP,
+            score=Common.DBotScore.GOOD
+        )
+
+        results = CommandResults(
+            outputs_key_field=None,
+            outputs_prefix=None,
+            outputs=None,
+            indicators=[dbot_score]
+        )
+
+        assert results.to_context() == {
+            'Type': EntryType.NOTE,
+            'ContentsFormat': EntryFormat.JSON,
+            'Contents': None,
+            'HumanReadable': None,
+            'EntryContext': {
+                'DBotScore(val.Indicator && val.Indicator == obj.Indicator && '
+                'val.Vendor == obj.Vendor && val.Type == obj.Type)': {
+                    'Indicator': '8.8.8.8',
+                    'Vendor': 'Virus Total',
+                    'Score': 1,
+                    'Type': 'ip'
+                }
+            }
+        }
+
+    def test_return_list_of_items(self):
+        from CommonServerPython import CommandResults, EntryFormat, EntryType
+        tickets = [
+            {
+                'ticket_id': 1,
+                'title': 'foo'
+            },
+            {
+                'ticket_id': 2,
+                'title': 'goo'
+            }
+        ]
+        results = CommandResults(
+            outputs_prefix='Jira.Ticket',
+            outputs_key_field='ticket_id',
+            outputs=tickets
+        )
+
+        assert results.to_context() == {
+            'Type': EntryType.NOTE,
+            'ContentsFormat': EntryFormat.JSON,
+            'Contents': tickets,
+            'HumanReadable': tableToMarkdown('Results', tickets),
+            'EntryContext': {
+                'Jira.Ticket(val.ticket_id == obj.ticket_id)': tickets
+            }
+        }
+
+    def test_create_dbot_score_with_invalid_score(self):
+        from CommonServerPython import Common, DBotScoreType
+
+        try:
+            Common.DBotScore(
+                indicator='8.8.8.8',
+                integration_name='Virus Total',
+                score=100,
+                indicator_type=DBotScoreType.IP
+            )
+
+            assert False
+        except TypeError:
+            assert True
+
+    def test_create_ip(self):
+        from CommonServerPython import Common
+
+        ip = Common.IP(
+            ip='8.8.8.8',
+            asn='some asn',
+            hostname='test.com',
+            geo_country=None,
+            geo_description=None,
+            geo_latitude=None,
+            geo_longitude=None,
+            positive_engines=None,
+            detection_engines=None
+        )
+
+        assert ip is not None
+
+    def test_create_domain(self):
+        from CommonServerPython import CommandResults, Common, EntryType, EntryFormat, DBotScoreType
+
+        domain = Common.Domain(
+            domain='somedomain.com',
+            dns='dns.somedomain',
+            detection_engines=10,
+            positive_detections=5,
+            organization='Some Organization',
+            whois=Common.WHOIS(
+                admin_phone='18000000',
+                admin_email='admin@test.com',
+
+                registrant_name='Mr Registrant',
+
+                registrar_name='Mr Registrar',
+                registrar_abuse_email='registrar@test.com'
+            ),
+            creation_date='2019-01-01T00:00:00',
+            update_date='2019-01-02T00:00:00',
+            expiration_date=None,
+            domain_status='ACTIVE',
+            name_servers=[
+                'PNS31.CLOUDNS.NET',
+                'PNS32.CLOUDNS.NET'
+            ],
+            sub_domains=[
+                'sub-domain1.somedomain.com',
+                'sub-domain2.somedomain.com',
+                'sub-domain3.somedomain.com'
+            ]
+        )
+
+        dbot_score = Common.DBotScore(
+            indicator='somedomain.com',
+            integration_name='Virus Total',
+            indicator_type=DBotScoreType.DOMAIN,
+            score=Common.DBotScore.GOOD
+        )
+        domain.set_dbot_score(dbot_score)
+
+        results = CommandResults(
+            outputs_key_field=None,
+            outputs_prefix=None,
+            outputs=None,
+            indicators=[domain]
+        )
+
+        assert results.to_context() == {
+            'Type': EntryType.NOTE,
+            'ContentsFormat': EntryFormat.JSON,
+            'Contents': None,
+            'HumanReadable': None,
+            'EntryContext': {
+                'Domain(val.Name && val.Name == obj.Name)': {
+                    'Name': 'somedomain.com',
+                    'DNS': 'dns.somedomain',
+                    'DetectionEngines': 10,
+                    'PositiveDetections': 5,
+                    'Organization': 'Some Organization',
+                    'CreationDate': '2019-01-01T00:00:00',
+                    'UpdateDate': '2019-01-02T00:00:00',
+                    'DomainStatus': 'ACTIVE',
+                    'NameServers': [
+                        'PNS31.CLOUDNS.NET',
+                        'PNS32.CLOUDNS.NET'
+                    ],
+                    'Subdomains': [
+                        'sub-domain1.somedomain.com',
+                        'sub-domain2.somedomain.com',
+                        'sub-domain3.somedomain.com'
+                    ],
+                    'Admin': {
+                        'Phone': '18000000',
+                        'Email': 'admin@test.com',
+                        'Name': None
+                    },
+                    'Registrant': {
+                        'Name': 'Mr Registrant',
+                        'Email': None,
+                        'Phone': None
+                    },
+                    'WHOIS': {
+                        'Admin': {
+                            'Name': None,
+                            'Phone': '18000000',
+                            'Email': 'admin@test.com'
+                        },
+                        'Registrar': {
+                            'Name': 'Mr Registrar',
+                            'AbuseEmail': 'registrar@test.com',
+                            'AbusePhone': None
+                        },
+                        'Registrant': {
+                            'Name': 'Mr Registrant',
+                            'Email': None,
+                            'Phone': None
+                        }
+                    }
+                },
+                'DBotScore(val.Indicator && val.Indicator == obj.Indicator && '
+                'val.Vendor == obj.Vendor && val.Type == obj.Type)': {
+                    'Indicator': 'somedomain.com',
+                    'Vendor': 'Virus Total',
+                    'Score': 1,
+                    'Type': 'domain'
+                }
+            }
+        }
+
+
 class TestBaseClient:
     from CommonServerPython import BaseClient
     text = {"status": "ok"}
@@ -916,6 +1202,28 @@ class TestBaseClient:
         requests_mock.get('http://example.com/api/v2/event', exc=requests.exceptions.ConnectionError)
         with raises(DemistoException, match="Verify that the server URL parameter"):
             self.client._http_request('get', 'event', resp_type='response')
+
+    def test_text_exception_parsing(self, requests_mock):
+        from CommonServerPython import DemistoException
+        reason = 'Bad Request'
+        text = 'additional text'
+        requests_mock.get('http://example.com/api/v2/event',
+                          status_code=400,
+                          reason=reason,
+                          text=text)
+        with raises(DemistoException, match='- {}\n{}'.format(reason, text)):
+            self.client._http_request('get', 'event', resp_type='text')
+
+    def test_json_exception_parsing(self, requests_mock):
+        from CommonServerPython import DemistoException
+        reason = 'Bad Request'
+        json_response = {'error': 'additional text'}
+        requests_mock.get('http://example.com/api/v2/event',
+                          status_code=400,
+                          reason=reason,
+                          json=json_response)
+        with raises(DemistoException, match='- {}\n.*{}'.format(reason, json_response["error"])):
+            self.client._http_request('get', 'event', resp_type='text')
 
     def test_is_valid_ok_codes_empty(self):
         from requests import Response
@@ -1207,3 +1515,37 @@ data_test_b64_encode = [
 def test_b64_encode(_input, expected_output):
     output = b64_encode(_input)
     assert output == expected_output, 'b64_encode({}) returns: {} instead: {}'.format(_input, output, expected_output)
+
+
+def test_traceback_in_return_error_debug_mode_on(mocker):
+    mocker.patch.object(demisto, 'command', return_value="test-command")
+    mocker.spy(demisto, 'results')
+    mocker.patch('CommonServerPython.is_debug_mode', return_value=True)
+    from CommonServerPython import return_error
+
+    try:
+        raise Exception("This is a test string")
+    except Exception:
+        with pytest.raises(SystemExit):
+            return_error("some text")
+
+    assert "This is a test string" in str(demisto.results.call_args)
+    assert "Traceback" in str(demisto.results.call_args)
+    assert "some text" in str(demisto.results.call_args)
+
+
+def test_traceback_in_return_error_debug_mode_off(mocker):
+    mocker.patch.object(demisto, 'command', return_value="test-command")
+    mocker.spy(demisto, 'results')
+    mocker.patch('CommonServerPython.is_debug_mode', return_value=False)
+    from CommonServerPython import return_error
+
+    try:
+        raise Exception("This is a test string")
+    except Exception:
+        with pytest.raises(SystemExit):
+            return_error("some text")
+
+    assert "This is a test string" not in str(demisto.results.call_args)
+    assert "Traceback" not in str(demisto.results.call_args)
+    assert "some text" in str(demisto.results.call_args)