+++ Packs/FeedTorExitAddresses/Integrations/FeedTorExitAddresses/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+Fixed an issue where the integration did not fetch indicators.
 
 ## [20.4.0] - 2020-04-14
 Added the *Tags* parameter.