+++ Packs/FeedTorExitAddresses/Integrations/FeedTorExitAddresses/CHANGELOG.md
@@ -1,6 +1,12 @@
 ## [Unreleased]
 
 
+## [20.4.1] - 2020-04-29
+Fixed an issue where the integration did not fetch indicators.
+
+## [20.4.0] - 2020-04-14
+Added the *Tags* parameter.
+
 ## [20.3.4] - 2020-03-30
 -